split db file into its own module + fixing db bugs + commenting code
