+++ src/db/mod.rs
@@ -1,0 +1,4 @@
+#![allow(dead_code)]
+
+pub mod summoner;
+pub mod summoner_rankeds;
